pyproject.toml
[project]
name = "omni-retail-de"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=40.13.0",
    "loguru>=0.7.3",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.3",
    "python-dotenv[cli]>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
]
